+++ src/contentsharecontroller/DefaultContentShareController.ts
@@ -9,6 +9,7 @@ import MeetingSessionConfiguration from '../meetingsession/MeetingSessionConfigu
 import MeetingSessionCredentials from '../meetingsession/MeetingSessionCredentials';
 import MeetingSessionStatus from '../meetingsession/MeetingSessionStatus';
 import AsyncScheduler from '../scheduler/AsyncScheduler';
+import VideoTile from '../videotile/VideoTile';
 import ContentShareConstants from './ContentShareConstants';
 import ContentShareController from './ContentShareController';
 import ContentShareMediaStreamBroker from './ContentShareMediaStreamBroker';
@@ -24,18 +25,21 @@ export default class DefaultContentShareController
     contentShareConfiguration.credentials = new MeetingSessionCredentials();
     contentShareConfiguration.credentials.attendeeId =
       configuration.credentials.attendeeId + ContentShareConstants.Modality;
+    contentShareConfiguration.credentials.externalUserId = configuration.credentials.externalUserId;
     contentShareConfiguration.credentials.joinToken =
       configuration.credentials.joinToken + ContentShareConstants.Modality;
     return contentShareConfiguration;
   }
 
   private observerQueue: Set<ContentShareObserver> = new Set<ContentShareObserver>();
+  private contentShareTile: VideoTile;
 
   constructor(
     private mediaStreamBroker: ContentShareMediaStreamBroker,
-    private audioVideo: AudioVideoController
+    private contentAudioVideo: AudioVideoController,
+    private attendeeAudioVideo: AudioVideoController
   ) {
-    this.audioVideo.addObserver(this);
+    this.contentAudioVideo.addObserver(this);
   }
 
   async startContentShare(stream: MediaStream): Promise<void> {
@@ -48,9 +52,9 @@ export default class DefaultContentShareController
         this.stopContentShare();
       });
     }
-    this.audioVideo.start();
+    this.contentAudioVideo.start();
     if (this.mediaStreamBroker.mediaStream.getVideoTracks().length > 0) {
-      this.audioVideo.videoTileController.startLocalVideoTile();
+      this.contentAudioVideo.videoTileController.startLocalVideoTile();
     }
   }
 
@@ -83,7 +87,7 @@ export default class DefaultContentShareController
   }
 
   stopContentShare(): void {
-    this.audioVideo.stop();
+    this.contentAudioVideo.stop();
     this.mediaStreamBroker.cleanup();
   }
 
@@ -106,6 +110,16 @@ export default class DefaultContentShareController
   }
 
   audioVideoDidStart(): void {
+    this.contentShareTile = this.attendeeAudioVideo.videoTileController.addVideoTile();
+    this.contentShareTile.bindVideoStream(
+      this.contentAudioVideo.configuration.credentials.attendeeId,
+      false,
+      this.mediaStreamBroker.mediaStream,
+      null,
+      null,
+      null,
+      this.contentAudioVideo.configuration.credentials.externalUserId
+    );
     this.forEachContentShareObserver(observer => {
       Maybe.of(observer.contentShareDidStart).map(f => f.bind(observer)());
     });
@@ -115,6 +129,10 @@ export default class DefaultContentShareController
     //If the content attendee got dropped or could not connect, stopContentShare will not be called
     //So make sure to clean up the media stream.
     this.mediaStreamBroker.cleanup();
+    if (this.contentShareTile) {
+      this.attendeeAudioVideo.videoTileController.removeVideoTile(this.contentShareTile.id());
+      this.contentShareTile = null;
+    }
     this.forEachContentShareObserver(observer => {
       Maybe.of(observer.contentShareDidStop).map(f => f.bind(observer)());
     });